Update LLM Functions
