+++ OwnYourData/TrialsMatching/MatchingState.swift
@@ -1,0 +1,16 @@
+//
+// This source file is part of the OwnYourData based on the Stanford Spezi Template Application project
+//
+// SPDX-FileCopyrightText: 2023 Stanford University
+//
+// SPDX-License-Identifier: MIT
+//
+
+import Foundation
+
+
+enum MatchingState {
+    case fhirInspection
+    case nciLoading
+    case matching
+}
