+++ OwnYourData/TrialsMatching/MatchingStateView.swift
@@ -1,0 +1,25 @@
+//
+//  MatchingStateView.swift
+//  OwnYourData
+//
+//  Created by Paul Shmiedmayer on 5/14/24.
+//
+
+import SwiftUI
+
+
+struct MatchingStateView: View {
+    @Environment(MatchingModule.self) private var matchingModule
+    
+    
+    var body: some View {
+        Group {
+            // ...
+        }
+    }
+}
+
+
+#Preview {
+    MatchingStateView()
+}
